+++ src/rateLimiter.ts
@@ -1,33 +1,43 @@
 export class RateLimiter {
   private queue: [() => Promise<unknown>, (x: unknown) => void, (x: unknown) => void][];
-  private timeout = 1000 / 2;
-  private timer: ReturnType<typeof setInterval> | null;
+  
+  private timeout1 = 1000 / 2; // Targeting 2 requests per second
+  private timeout2 = 1000 / (30 / 60); // Targeting 30 requests per minute
+  
+  private timer1: ReturnType<typeof setInterval> | null;
+  private timer2: ReturnType<typeof setInterval> | null;
+  private timer2OffsetTimer: ReturnType<typeof setTimeout> | null;
 
   constructor() {
     this.queue = [];
-    this.timer = null;
+  
+    this.timer1 = null;
+    this.timer2 = null;
+    this.timer2OffsetTimer = null;
   }
 
   enqueue<T>(fn: () => Promise<T>): Promise<T> {
-    return new Promise((resolve, reject) => {
-      this.queue.push([fn, resolve, reject]);
-      setTimeout(() => this.ensureWorking(), 0);
-    });
+    setTimeout(() => this.ensureWorking(), 0); // Next tick
+    return new Promise((resolve, reject) => this.queue.push([fn, resolve, reject]));
+  }
+
+  // Put an item for a retry at the front of the queue
+  retry<T>(fn: () => Promise<T>): Promise<T> {
+    setTimeout(() => this.ensureWorking(), 0); // Next tick
+    return new Promise((resolve, reject) => this.queue.unshift([fn, resolve, reject]));
   }
 
   private ensureWorking() {
-    if (this.timer != null) {
+    if (this.timer1 != null) {
       return;
     }
 
-    this.timer = setInterval(
-      () => {
-        if (!this.work()) {
-          this.stop();
-        }
-      },
-      this.timeout,
-    );
+    this.timer1 = setInterval(() => this.work(), this.timeout1);
+
+    // Start the second timer in between the first and second request, weaving in the burst request between the normal requests
+    this.timer2OffsetTimer = setTimeout(() => {
+      this.timer2 = setInterval(() => this.work(), this.timeout2);
+    }, this.timeout1 / 2);
 
     this.work();
   }
@@ -36,7 +46,8 @@ export class RateLimiter {
     const item = this.queue.shift();
   
     if (!item) {
-      return false;
+      this.stop();
+      return;
     }
 
     const [fn, resolve, reject] = item;
@@ -44,12 +55,15 @@ export class RateLimiter {
     fn()
       .then(resolve)
       .catch(reject);
-    
-    return true;
   }
 
   private stop() {
-    clearInterval(this.timer);
-    this.timer = null;
+    clearInterval(this.timer1);
+    clearInterval(this.timer2);
+    clearTimeout(this.timer2OffsetTimer);
+  
+    this.timer1 = null;
+    this.timer2 = null;
+    this.timer2OffsetTimer = null;
   }
 }